Change . to a - in filename as expected

+++ CI/appveyor.after_success.ps1
@@ -58,15 +58,18 @@ if ("$Env:APPVEYOR_REPO_TAG" -eq "false" -and -Not $public_test_build) {
 
   Write-Output "=== Creating installers from Nuget package ==="
   if ($public_test_build) {
-    $TestBuildString = ".PublicTestBuild"
+    $TestBuildString = "-PublicTestBuild"
   } else {
     $TestBuildString = ""
   }
-  .\squirrel.windows\tools\Squirrel --releasify C:\projects\squirrel-packaging-prep\Mudlet$($TestBuildString).$($Env:VERSION).nupkg --releaseDir C:\projects\squirreloutput --loadingGif C:\projects\installers\windows\splash-installing-2x.png --no-msi --setupIcon C:\projects\installers\windows\mudlet_main_48px.ico -n "/a /f C:\projects\installers\windows\code-signing-certificate.p12 /p $Env:signing_password /fd sha256 /tr http://timestamp.digicert.com /td sha256"
 
-  Write-Output "=== Printing contents of $Env:APPVEYOR_BUILD_FOLDER\src ==="
-  Tree /F $Env:APPVEYOR_BUILD_FOLDER\src
-  Write-Output "=== Done printing ==="
+  if (-not (Test-Path -Path C:\projects\squirrel-packaging-prep\Mudlet$($TestBuildString).$($Env:VERSION).nupkg -PathType Leaf)) {
+    Write-Output "=== ERROR: nupkg doesn't exist as expected! Build aborted."
+    exit 1
+  }
+
+  # fails silently if the file is not found
+  .\squirrel.windows\tools\Squirrel --releasify C:\projects\squirrel-packaging-prep\Mudlet$($TestBuildString).$($Env:VERSION).nupkg --releaseDir C:\projects\squirreloutput --loadingGif C:\projects\installers\windows\splash-installing-2x.png --no-msi --setupIcon C:\projects\installers\windows\mudlet_main_48px.ico -n "/a /f C:\projects\installers\windows\code-signing-certificate.p12 /p $Env:signing_password /fd sha256 /tr http://timestamp.digicert.com /td sha256"
 
   Write-Output "=== Removing old directory content of release folder ==="
   Remove-Item -Recurse -Force $Env:APPVEYOR_BUILD_FOLDER\src\release\*
